% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reed_thesis.R
\name{label}
\alias{label}
\title{Adds a label to a figure/chemical reaction and includes the figure in the
document}
\usage{
label(path = NULL, caption = "Default caption", label = "def",
  type = "figure", alt.cap = caption, cap.size = "normalsize",
  scale = 1, angle = 0, options = "h!tbp")
}
\arguments{
\item{path}{A string specifying where the object is stored as a file}

\item{caption}{A string providing the text that will appear with the object}

\item{label}{A string providing a name to use with \code{ref}}

\item{type}{A string (either "figure" or "chem.reac") providing object type}

\item{alt.cap}{A string giving an alternative caption to be included in the
Table of Figures}

\item{cap.size}{A string giving the size of the caption.  Options are
"scriptsize", "footnotesize", "small", "normalsize", "large", and "Large"}

\item{scale}{A numeric specifying the scale to display a figure (1
corresponds to 100\% of original size, 0.5 corresponds to 50\% of original
size, etc.)}

\item{angle}{A numeric specifying the degree to rotate a figure}

\item{options}{A string attempting to fix some of the "floating" issues with
LaTeX images}
}
\value{
LaTeX figure environment code that can be run with R chunk option
  \code{results="asis"} to produce object in document with given label and
  features
}
\description{
This allows users to label and include R and non-R generated objects into
their document
}
\examples{
\dontrun{
 label(path = "figure/reed.jpg", caption = "Reed logo",
       label = "reed", type = "figure")
}
}

