% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reed_thesis.R
\name{reed_thesis}
\alias{reed_thesis}
\title{Creates an R Markdown PDF Thesis document}
\usage{
reed_thesis(toc = TRUE, toc_depth = 3)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying where table of contents should be created}

\item{toc_depth}{A positive integer}
}
\value{
A modified \code{pdf_document} based on the Reed Senior Thesis LaTeX
  template
}
\description{
This is a function called in output in the YAML of the driver Rmd file to specify using the Reed
Senior Thesis LaTeX template and cls files.
}
\examples{
\dontrun{
 output:
   reedtemplates::reed_thesis:
     toc: true
}
}

