% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reedtemplates.R
\docType{package}
\name{reedtemplates}
\alias{reedtemplates}
\alias{reedtemplates-package}
\title{reedtemplates: A package for creating documents using R Markdown}
\description{
The reedtemplates package provides three important functions:
reed_thesis, label, and ref.
}
\section{reed_thesis}{

Creates an R Markdown senior thesis template
}

\section{label}{

Adds a label to a figure/chemical reaction and includes the figure in the document
}

\section{ref}{

Allows references to be made to labelled objects inline in the document
}

