% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reed_thesis.R
\name{ref}
\alias{ref}
\title{References previously created label}
\usage{
ref(label = "def", type = "figure")
}
\arguments{
\item{label}{A string specifying object label given in \code{label} function
call}

\item{type}{A string corresponding to "table" or "figure" label}
}
\value{
LaTeX code to provide object number corresponding to \code{label}
  specified in \code{label} function
}
\description{
This allows users to make a reference to an object imported using the
\code{label} function
}
\examples{
\dontrun{
 `r ref("subd2", type = "figure")`
}
}

